[REFACTOR] addFriend에서 addEditFriend로 네이밍 변경
