+++ app/src/main/java/com/ivyclub/contact/ui/main/add_edit_friend/FriendExtraInfoData.kt
@@ -1,4 +1,4 @@
-package com.ivyclub.contact.ui.main.add_friend
+package com.ivyclub.contact.ui.main.add_edit_friend
 
 data class FriendExtraInfoData(var title: String, var value: String) {
 }
