+++ app/src/main/res/layout/item_add_friend_extra_info.xml
@@ -16,8 +16,8 @@
             android:id="@+id/et_extra_info_title"
             android:layout_width="wrap_content"
             android:layout_height="20dp"
-            android:background="@drawable/bg_add_friend_info"
-            android:hint="@string/add_extra_info_i_title_hint"
+            android:background="@drawable/bg_add_edit_friend_info"
+            android:hint="@string/add_edit_friend_extra_info_title_hint"
             android:paddingHorizontal="8dp"
             android:textSize="12sp"
             android:textStyle="bold"
@@ -40,7 +40,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
-            android:background="@drawable/bg_add_friend_info"
+            android:background="@drawable/bg_add_edit_friend_info"
             android:paddingHorizontal="8dp"
             android:paddingVertical="12dp"
             android:textSize="16sp"
